+++ config_wizard.ps1
@@ -297,12 +297,11 @@ $Config = [ordered]@{
     cloud_folder_name = $CloudFolderName
     remote_root = $RemoteRoot
     cleanup_mode = $CleanupMode
-    task_name = 'ScreenAgent-AutoUpload'
     scan_interval_seconds = 15
     stable_seconds = 12
     cloud_type = $CloudType
     shortcut_name = $ShortcutName
-    config_version = 3
+    config_version = 4
 }
 
 $Utf8Bom = New-Object System.Text.UTF8Encoding($true)